\name{findNamedFunctions}
\alias{findNamedFunctions}
\title{Find Named/Non-Anonymous Nested Functions}
\description{
Given a function, find the named-function definitions.
}
\usage{
findNamedFunctions(fun)
}
\arguments{
  \item{fun}{a function}
}
\value{
  A list of assignment calls in which the
  right hand side is a call to \code{function}.
}
\author{Duncan Temple Lang}

%\seealso{}
%\examples{}

\keyword{programming}
\concept{meta-programming}

