\name{findCallsParam}
\alias{findCallsParam}
\title{Determine if a function directly invokes any its arguments as a function}
\description{
  This function analyzes the code in a function and determines
  if it directly or indirectly invokes any of its parameters
  and, if so, which ones.
}
\usage{
findCallsParam(fun)
}
\arguments{
  \item{fun}{the R function}
}
\value{
}
\author{
Duncan Temple Lang
}
\seealso{
}
\examples{
}
\keyword{programming}
\concept{static code analysis}
